+++ src/RcppExports.cpp
@@ -223,18 +223,6 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
-// bias
-NumericVector bias(NumericMatrix dvs, int nd);
-RcppExport SEXP _redistmetrics_bias(SEXP dvsSEXP, SEXP ndSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< NumericMatrix >::type dvs(dvsSEXP);
-    Rcpp::traits::input_parameter< int >::type nd(ndSEXP);
-    rcpp_result_gen = Rcpp::wrap(bias(dvs, nd));
-    return rcpp_result_gen;
-END_RCPP
-}
 // declination_simple
 NumericVector declination_simple(NumericMatrix dvs, IntegerVector dseat_vec, int nd);
 RcppExport SEXP _redistmetrics_declination_simple(SEXP dvsSEXP, SEXP dseat_vecSEXP, SEXP ndSEXP) {
@@ -321,6 +309,18 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< NumericMatrix >::type dvs(dvsSEXP);
     Rcpp::traits::input_parameter< int >::type nd(ndSEXP);
     rcpp_result_gen = Rcpp::wrap(smoothseat(dvs, nd));
+    return rcpp_result_gen;
+END_RCPP
+}
+// reindex
+IntegerMatrix reindex(IntegerMatrix dm, int nd);
+RcppExport SEXP _redistmetrics_reindex(SEXP dmSEXP, SEXP ndSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< IntegerMatrix >::type dm(dmSEXP);
+    Rcpp::traits::input_parameter< int >::type nd(ndSEXP);
+    rcpp_result_gen = Rcpp::wrap(reindex(dm, nd));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -420,7 +420,6 @@ static const R_CallMethodDef CallEntries[] = {
     {"_redistmetrics_effgap", (DL_FUNC) &_redistmetrics_effgap, 3},
     {"_redistmetrics_taugap", (DL_FUNC) &_redistmetrics_taugap, 4},
     {"_redistmetrics_meanmedian", (DL_FUNC) &_redistmetrics_meanmedian, 1},
-    {"_redistmetrics_bias", (DL_FUNC) &_redistmetrics_bias, 2},
     {"_redistmetrics_declination_simple", (DL_FUNC) &_redistmetrics_declination_simple, 3},
     {"_redistmetrics_declination_angle", (DL_FUNC) &_redistmetrics_declination_angle, 3},
     {"_redistmetrics_lopsidedwins", (DL_FUNC) &_redistmetrics_lopsidedwins, 3},
@@ -428,6 +427,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_redistmetrics_biasatv", (DL_FUNC) &_redistmetrics_biasatv, 3},
     {"_redistmetrics_RankedMarginalDev", (DL_FUNC) &_redistmetrics_RankedMarginalDev, 1},
     {"_redistmetrics_smoothseat", (DL_FUNC) &_redistmetrics_smoothseat, 2},
+    {"_redistmetrics_reindex", (DL_FUNC) &_redistmetrics_reindex, 2},
     {"_redistmetrics_segregationcalc", (DL_FUNC) &_redistmetrics_segregationcalc, 3},
     {"_redistmetrics_splits", (DL_FUNC) &_redistmetrics_splits, 4},
     {"_redistmetrics_distr_cty_splits", (DL_FUNC) &_redistmetrics_distr_cty_splits, 3},