+++ NEWS.md
@@ -1,3 +1,9 @@
+# redistmetrics 1.0.1
+
+* Fixed bug where `part_bias()` was reversed in sign.
+* Fixed bug where `splits_admin()` plans input needed to be sequentially numbered (#10).
+* Allows planarizing warnings to be silenced when `epsg = FALSE` (#11).
+
 # redistmetrics 0.0.0.9000
 
 * Added a `NEWS.md` file to track changes to the package.